wip(quantization): Demote QuantizerType, used only in tests, to the test header file.
